+++ app/src/main/java/com/example/budgetcontrol/db/model/GaussNumber.kt
@@ -6,9 +6,9 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "gauss_numbers")
 data class GaussNumber(
-    @PrimaryKey(autoGenerate = true) val id: Int,
-    val value: Float,
-    @ColumnInfo(name = "is_collected") val isCollected: Boolean
+        @PrimaryKey(autoGenerate = true) val id: Int,
+        val value: Int,
+        @ColumnInfo(name = "is_collected") val isCollected: Boolean
 ){
-    constructor(value: Float, isCollected: Boolean) : this(0, value, isCollected)
+    constructor(value: Int, isCollected: Boolean) : this(0, value, isCollected)
 }
